Rename experimental annotations
