+++ dlup/data/experimental/dataset.py
@@ -8,8 +8,8 @@ from typing import Callable, Iterable, List, Optional, Tuple, Union
 import numpy as np
 
 from dlup import SlideImage
+from dlup.annotations import WsiAnnotations
 from dlup.data.dataset import TiledROIsSlideImageDataset, parse_rois
-from dlup.experimental_annotations import WsiAnnotations
 from dlup.experimental_backends import ImageBackends
 from dlup.tiling import Grid, GridOrder, TilingMode
